+++ openfisca_core/taxscales/rate_tax_scale_like.py
@@ -14,7 +14,7 @@ from openfisca_core.errors import EmptyArgumentError
 from .tax_scale_like import TaxScaleLike
 
 if typing.TYPE_CHECKING:
-    NumericalArray = typing.Union[numpy.int_, numpy.float64]
+    NumericalArray = typing.Union[numpy.int32, numpy.float32]
 
 
 class RateTaxScaleLike(TaxScaleLike, abc.ABC):
@@ -128,7 +128,7 @@ class RateTaxScaleLike(TaxScaleLike, abc.ABC):
         tax_base: NumericalArray,
         factor: float = 1.0,
         round_decimals: int | None = None,
-    ) -> numpy.int_:
+    ) -> numpy.int32:
         """Compute the relevant bracket indices for the given tax bases.
 
         :param ndarray tax_base: Array of the tax bases.
